+++ index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="public/css/style.css">
+    <link rel="stylesheet" href="/public/css/style.css">
     <link rel="manifest" href="manifest.json">
     <title>Canvas</title>
 </head>
@@ -109,8 +109,8 @@
         </main>
     </div>
     <dialog is="input-modal"></dialog>
-    <script src="./public/components/InputModal.js" type="module"></script>
-    <script src="./public/application/app.js" type="module"></script>
+    <script src="/src/components/InputModal.js" type="module"></script>
+    <script src="/src/application/app.js" type="module"></script>
 </body>
 
 </html>
