+++ public/canvas.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="css/canvas.css">
+    <link rel="stylesheet" href="./css/canvas.css">
     <title>Editando desenho</title>
 </head>
 
@@ -68,7 +68,7 @@
             <input id="drawing-line-width" type="number" min="2" value="6">
         </div>
     </div>
-    <script src="application/canvas.js" type="module"></script>
+    <script src="/src/application/drawing/canvas.js" type="module"></script>
 </body>
 
 </html>
