+++ src/application/drawing/canvas.js
@@ -19,450 +19,9 @@ canvas.onmouseup = () => {
 }
 */
 
-import SharedWorkerPolyfill from "../polyfill/SharedWorkerPolyfill.js";
-import toArray from "../utils/toArray.js";
-
-class Draw {
-    /**
-     * @param {HTMLCanvasElement} canvas  canvas para o qual será desenhado.
-    */
-    constructor(canvas) {
-        this.canvas = canvas;
-        this.ctx = canvas.getContext("2d");
-        this.fill = "black";
-    }
-
-    export() {
-        let url;
-        const a = document.createElement("a");
-        this.canvas.toBlob((blob) => {
-            a.href = URL.createObjectURL(blob);
-            url = a.href;
-            a.download = "desenho";
-            a.click();
-            setTimeout(() => {
-                URL.revokeObjectURL(url);
-                a.remove();
-            }, 50);
-        })
-    }
-
-    clear() {
-        this.ctx.save();
-        this.ctx.resetTransform();
-        this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
-        this.ctx.restore();
-    }
-
-    /**
-     * Move the drawing pen to the especified coords.
-     * @param {Number} x position in the x-axis in pixels.
-     * @param {Number} y position in the y-axis in pixels.
-     */
-    moveTo(x, y) {
-        this.ctx.moveTo(x, y);
-    }
-
-    /**
-     * Create a line from the point [x1, y1] to the point [x2, y2].
-     * @param {Number} x1 first point position in the x-axis in pixels. 
-     * @param {Number} y1 first point position in the y-axis in pixels.
-     * @param {Number} x2 second point position in the x-axis in pixels.
-     * @param {Number} y2 second point position in the y-axis in pixels.
-     */
-    line(x1, y1, x2, y2) {
-        this.ctx.moveTo(x1, y1);
-        this.ctx.lineTo(x2, y2);
-    }
-
-    /**
-     * Create a line from the point [x1, y1] to the point [x2, y2] and show in the canvas.
-     * @param {Number} x1 first point position in the x-axis in pixels.
-     * @param {Number} y1 first point position in the y-axis in pixels.
-     * @param {Number} x2 second point position in the x-axis in pixels.
-     * @param {Number} y2 second point position in the y-axis in pixels.
-     */
-    strokeLine(x1, y1, x2, y2) {
-        this.ctx.moveTo(x1, y1);
-        this.ctx.lineTo(x2, y2);
-        this.ctx.stroke();
-    }
-
-    /**
-     * Draw a line to the point [x, y].
-     * @param {Number} x where the line will end in the x-axis in pixels
-     * @param {Number} y where the line will end in the y-axis in pixels
-     */
-    lineTo(x, y) {
-        this.ctx.lineTo(x, y);
-    }
-
-    /**
-     * Draw a line to the point [x, y] and show in the canvas.
-     * @param {Number} x where the line will end in the x-axis in pixels.
-     * @param {Number} y where the line will end in the y-axis in pixels.
-     */
-    strokeLineTo(x, y) {
-        this.ctx.lineTo(x, y);
-        this.ctx.stroke();
-    }
-
-    /**
-     * Draw a circle centered in the points x and y.
-     * @param {Number} x center of the circle position in the x-axis in pixels.
-     * @param {Number} y center of the circle position in the y-axis in pixels.
-     * @param {Number} radius half the diameter of the circle in pixels.
-     * @param {Number} startAngle initial angle in radians.
-     * @param {Boolean} isFilled boolean value that indicates if the circle will be filled after being drawn.
-     */
-    circle(x, y, radius, startAngle, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + radius, y);
-        this.ctx.arc(x, y, Math.abs(radius), startAngle, 2 * Math.PI);
-
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a rectangle.
-     * @param {Number} x rectangle position in the x-axis in pixels.
-     * @param {Number} y rectangle position in the y-axis in pixels.
-     * @param {Number} w rectangle width in pixels.
-     * @param {Number} h rectangle height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the rectangle will be filled after being drawn.
-    */
-    rectangle(x, y, w, h, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.rect(x, y, w, h);
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a diamond.
-     * @param {Number} x diamond position in the x-axis in pixels.
-     * @param {Number} y diamond position in the y-axis in pixels.
-     * @param {Number} w diamond width in pixels.
-     * @param {Number} h diamond height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the diamond will be filled after being drawn.
-    */
-    diamond(x, y, w, h, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x, y + h / 2);
-        this.ctx.lineTo(x + w / 2, y + h);
-        this.ctx.lineTo(x + w, y + h / 2);
-        this.ctx.lineTo(x + w / 2, y);
-
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a triangle.
-     * @param {Number} x triangle position in the x-axis in pixels.
-     * @param {Number} y triangle position in the y-axis in pixels.
-     * @param {Number} w triangle width in pixels.
-     * @param {Number} h triangle height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the triangle will be filled after being drawn.
-     */
-    triangle(x, y, w, h, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x, y + h);
-        this.ctx.lineTo(x + w, y + h);
-        this.ctx.lineTo(x + w / 2, y);
-
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a equilateral triangle.
-     * @param {Number} x equilateral triangle position in the x-axis in pixels.
-     * @param {Number} y equilateral triangle position in the y-axis in pixels.
-     * @param {Number} size equilateral triangle size in pixels. Will be used as the width and height.
-     * @param {Boolean} isFilled boolean value that indicates if the equilateral triangle will be filled after being drawn.
-    */
-    equilateralTriangle(x, y, size, isFilled = false) {
-        let w = size;
-        let h = size * 0.8;
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x, y + h);
-        this.ctx.lineTo(x + w, y + h);
-        this.ctx.lineTo(x + w / 2, y);
-
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a star.
-     * @param {Number} x star position in the x-axis in pixels.
-     * @param {Number} y star position in the y-axis in pixels.
-     * @param {Number} w star width in pixels.
-     * @param {Number} h star height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the star will be filled after being drawn.
-     */
-    star(x, y, w, h, isFilled = false) {
-        let heightPart = h / 5;
-        let widthPart = w / 5;
-
-        this.ctx.beginPath();
-        // --------- "/" of the top cone (^). ---------
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x + widthPart * 1.85, y + heightPart * 1.8);
-        // --------- "\" of the top cone (^). ---------
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x + widthPart * 3.15, y + heightPart * 1.8);
-
-        // --------- top line of the left cone ("<"). ---------
-        // move to the lower left tip of the upper cone.
-        this.ctx.moveTo(x + widthPart * 1.85, y + heightPart * 1.8);
-        this.ctx.lineTo(x, y + heightPart * 1.8);
-        // --------- bottom line of the left cone ("<"). ---------
-        this.ctx.lineTo(x + w / 3 - (w / 18), y + (h / 2.7) * 1.667);
-
-        // --------- "\" of the left bottom cone. ---------
-        this.ctx.lineTo(x + (w / 3) / 2, y + h);
-        // --------- "/" of the left bottom cone. ---------
-        this.ctx.lineTo(x + w / 2, y + heightPart * 3.5);
-
-        // --------- "\" of the right bottom cone. ---------
-        this.ctx.lineTo(x + (w / 3) * 2.5, y + h);
-        // --------- "/" of the right bottom cone. ---------
-        this.ctx.lineTo(x + (w / 3) * 2.3 - (w / 18), y + (h / 2.7) * 1.667);
-
-        // --------- bottom line of the right cone (">"). ---------
-        this.ctx.lineTo(x + w, y + heightPart * 1.8);
-        // --------- top line of the right cone (">"). ---------
-        this.ctx.moveTo(x + widthPart * 3.15, y + heightPart * 1.8);
-        this.ctx.lineTo(x + w, y + heightPart * 1.8);
-
-        // show the drawing.
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a trapezoid.
-     * @param {Number} x trapezoid position in the x-axis in pixels.
-     * @param {Number} y trapezoid position in the x-axis in pixels.
-     * @param {Number} w trapezoid width in pixels.
-     * @param {Number} h trapezoid height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the trapezoid will be filled after being drawn.
-     */
-    trapezoid(x, y, w, h, isFilled = false) {
-        let widthPart = w / 5;
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + widthPart, y);
-        this.ctx.lineTo(x + w - widthPart, y);
-        this.ctx.lineTo(x + w, y + h);
-        this.ctx.lineTo(x, y + h);
-        this.ctx.lineTo(x + widthPart, y);
-
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a ellipse.
-     * @param {Number} x ellipse position in the x-axis in pixels.
-     * @param {Number} y ellipse position in the y-axis in pixels.
-     * @param {Number} w ellipse width in pixels.
-     * @param {Number} h ellipse height in pixels.
-     * @param {Number} rotation oval rotation in degrees.
-     * @param {Boolean} isFilled boolean value that indicates if the ellipse will be filled after being drawn.
-     */
-    ellipse(x, y, w, h, rotation = 0, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.ellipse(x, y, Math.abs(w), Math.abs(h), Math.abs(rotation) * Math.PI / 180, 0, 2 * Math.PI);
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a pentagon.
-     * @param {Number} x pentagon position in the x-axis in pixels.
-     * @param {Number} y pentagon position in the y-axis in pixels.
-     * @param {Number} w pentagon width in pixels.
-     * @param {Number} h pentagon height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the pentagon will be filled after being drawn.
-     */
-    pentagon(x, y, w, h, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + w / 2, y);
-        this.ctx.lineTo(x, y + h / 3 * 1.3);
-        this.ctx.lineTo(x + (w / 5) * 1, y + h);
-        this.ctx.lineTo(x + (w / 5) * 4, y + h);
-        this.ctx.lineTo(x + w, y + h / 3 * 1.3);
-        this.ctx.lineTo(x + w / 2, y);
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-    }
-
-    /**
-     * Draw a hexagon.
-     * @param {Number} x hexagon position in the x-axis in pixels.
-     * @param {Number} y hexagon position in the y-axis in pixels.
-     * @param {Number} w hexagon width in pixels.
-     * @param {Number} h hexagon height in pixels.
-     * @param {Boolean} isFilled boolean value that indicates if the hexagon will be filled after being drawn.
-    */
-    hexagon(x, y, w, h, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.moveTo(x + w / 8 * 2, y);
-        this.ctx.lineTo(x + w / 8 * 6, y);
-        this.ctx.lineTo(x + w, y + h / 2);
-        this.ctx.lineTo(x + w / 8 * 6, y + h);
-        this.ctx.lineTo(x + w / 8 * 2, y + h);
-        this.ctx.lineTo(x, y + h / 2);
-        this.ctx.lineTo(x + w / 8 * 2, y);
-        if (isFilled) {
-            this.ctx.fill();
-        } else {
-            this.ctx.stroke();
-        }
-
-        this.ctx.closePath();
-
-    }
-
-    /**
-     * Draw a rectangle with rounded borders.
-     * @param {Number} x position of the rectangle in the x-axis in pixels.
-     * @param {Number} y position of the rectangle in the y-axis in pixels.
-     * @param {Number} w width of the rectangle in pixels.
-     * @param {Number} h height of the rectangle in pixels.
-     * @param {Array} radii array of the borders radii in pixels.
-     *                      - The first value will be used to indicate the radii of the top left corner, the second value to the top right corner and so on.
-     *                      - If there is only one value, it will be used on all the borders.
-     *                      - If a corner does not have a value in the array, it will not be rounded.
-     * @param {Boolean} isFilled boolean value that indicates if the rectangle will be filled after being drawn.
-    */
-    roundRectangle(x, y, w, h, radii, isFilled = false) {
-        this.ctx.beginPath();
-
-        this.ctx.roundRect(x, y, w, h, radii);
-        this.ctx.stroke();
-
-        this.ctx.closePath();
-
-        if (isFilled) {
-            this.ctx.fill();
-        }
-    }
-
-    /**
-     * Draws a polygon with the amount of sides especified
-     * @param {Number} x position of the polygon in the x-axis in pixels.
-     * @param {Number} y position of the polygon in the y-axis in pixels.
-     * @param {Number} w width of the polygon in pixels.
-     * @param {Number} h height of the polygon in pixels.
-     * @param {Number} sides number of sides of the polygon.
-     */
-    polygon(x, y, w, h, sides) {
-        function toRad(deg) {
-            return deg / 180 * Math.PI;
-        }
-
-        const STEP = 360 / sides;
-        this.ctx.beginPath();
-
-        for (let i = 0; i < sides + 1; i++) {
-            // h and w will be the (y and x) radius of the circle (ellipse), what means that
-            // the (x and y) diameter of the circle will be h*2 and w*2 (respectively), then,
-            // we divide it by 2. For example, h, is the height of the polygon, so the diameter
-            // would be twice the size we want and because of that we only multiply by half h.
-            let ray = { dirX: 0, dirY: 0 };
-            if (sides % 2 === 1) {
-                ray.dirY = Math.sin(toRad(i * STEP + (STEP / 4))) * h / 2;
-                ray.dirX = Math.cos(toRad(i * STEP + (STEP / 4))) * w / 2;
-            } else {
-                ray.dirY = Math.sin(toRad(i * STEP)) * h / 2;
-                ray.dirX = Math.cos(toRad(i * STEP)) * w / 2;
-            }
-
-            // offset the x and y to be at the middle of the box the polygon should be.
-            this.ctx.lineTo(ray.dirX + x + (w / 2), ray.dirY + y + (h / 2));
-        }
-
-        this.ctx.stroke();
-    }
-
-    text(text, x, y, options = {}) {
-        this.ctx.save();
-        if (options.color) {
-            this.ctx.fillStyle = options.color;
-        }
-
-        this.ctx.font = `${options.fontSize ? options.fontSize : 10}px ${options.fontFamily ? options.fontFamily : "serif"}`;
-
-        if (options.maxWidth) {
-            this.ctx.fillText(text, x, y - parseFloat(this.ctx.font.split("px")[0].split(" ")[this.ctx.font.split("px")[0].split(" ").length - 1]), options.maxWidth);
-        } else {
-            this.ctx.fillText(text, x, y + parseFloat(this.ctx.font.split("px")[0].split(" ")[this.ctx.font.split("px")[0].split(" ").length - 1]));
-        }
-        this.ctx.restore();
-    }
-}
-
+import SharedWorkerPolyfill from "./../../lib/polyfill/SharedWorkerPolyfill.js";
+import toArray from "./../../utils/toArray.js";
+import Drawer from "./Drawer.js";
 
 const canvas = document.querySelector("#main-canvas");
 canvas.height = window.innerHeight;
@@ -472,7 +31,7 @@ const ctx = canvas.getContext("2d");
 const drawingPreview = document.querySelector("#drawing-preview");
 drawingPreview.height = canvas.height;
 drawingPreview.width = canvas.width;
-const previewDraw = new Draw(drawingPreview);
+const previewDraw = new Drawer(drawingPreview);
 
 const lineWidthInput = document.querySelector("input#drawing-line-width");
 const drawingColorInput = document.querySelector("input#drawing-line-color");
@@ -480,7 +39,7 @@ const resizer = document.querySelector("#resizer");
 const drawingOptions = document.querySelector("#drawing-options");
 const carac = document.querySelector("#caracteristics");
 const closeCanvasBtn = document.querySelector("#return");
-// if the user is drawing, not if the drawing is drawing.
+// if the user is drawing, not if the drawing is a drawing.
 let isDrawing = false;
 let lowestPosition = { x: 9999, y: 9999 };
 let highestPosition = { x: 0, y: 0 };
@@ -498,9 +57,9 @@ ctx.lineJoin = "round";
 // canvas.
 let isPixelSharpnessTranslated = false;
 
-const draw = new Draw(canvas);
+const drawer = new Drawer(canvas);
 
-const sharedWorker = new SharedWorkerPolyfill("../application/sharedWorker.js");
+const sharedWorker = new SharedWorkerPolyfill("./../../services/sharedWorker.js");
 
 sharedWorker.port.onmessage = function (ev) {
     const msg = ev.data;
@@ -512,7 +71,7 @@ sharedWorker.port.onmessage = function (ev) {
             img.src = URL.createObjectURL(msg.img);
             img.onload = function () {
                 URL.revokeObjectURL(img.src);
-                draw.ctx.drawImage(img, 0, 0);
+                drawer.ctx.drawImage(img, 0, 0);
             }
         } else {
             alert(msg.errorMsg);
@@ -527,9 +86,9 @@ sharedWorker.port.onmessage = function (ev) {
 }
 
 function init() {
-    draw.ctx.lineWidth = parseInt(carac.children["drawing-line-width"].value);
-    draw.ctx.strokeStyle = carac.children["drawing-line-color"].value;
-    draw.ctx.fillStyle = carac.children["drawing-line-color"].value;
+    drawer.ctx.lineWidth = parseInt(carac.children["drawing-line-width"].value);
+    drawer.ctx.strokeStyle = carac.children["drawing-line-color"].value;
+    drawer.ctx.fillStyle = carac.children["drawing-line-color"].value;
 }
 init();
 
@@ -585,9 +144,9 @@ function showTextCaracteristics() {
         textInputWrraper.appendChild(textInput);
 
 
-        fontSizeInput.oninput = function () { draw.ctx.font = `bold ${this.value}px ${fontFamilyInput.value}`; }
-        fontFamilyInput.oninput = function () { draw.ctx.font = `bold ${fontSizeInput.value}px ${this.value}`; }
-        drawingColorInput.oninput = function () { draw.ctx.fillStyle = this.value; }
+        fontSizeInput.oninput = function () { drawer.ctx.font = `bold ${this.value}px ${fontFamilyInput.value}`; }
+        fontFamilyInput.oninput = function () { drawer.ctx.font = `bold ${fontSizeInput.value}px ${this.value}`; }
+        drawingColorInput.oninput = function () { drawer.ctx.fillStyle = this.value; }
 
         if (fontsDatalist.children.length === 0) {
             (async function () {
@@ -662,7 +221,7 @@ function showNewPathButton() {
         const button = document.createElement("button");
         button.style.display = "block";
         button.textContent = "Iniciar novo traço";
-        button.addEventListener("click", function () { draw.ctx.beginPath() });
+        button.addEventListener("click", function () { drawer.ctx.beginPath() });
 
         carac.appendChild(button);
     }
@@ -729,8 +288,8 @@ canvas.addEventListener("touchcancel", function (ev) {
 
 canvas.addEventListener("click", function (ev) {
     if (currentDrawingMode === "line") {
-        draw.lineTo(Math.round(ev.offsetX), Math.round(ev.offsetY));
-        draw.ctx.stroke();
+        drawer.lineTo(Math.round(ev.offsetX), Math.round(ev.offsetY));
+        drawer.ctx.stroke();
     }
 });
 
@@ -741,10 +300,10 @@ function handleMouseOrTouchMove(event) {
 
     switch (currentDrawingMode) {
         case "free":
-            draw.strokeLineTo(getEventPos(event).x, getEventPos(event).y);
+            drawer.strokeLineTo(getEventPos(event).x, getEventPos(event).y);
             break;
         case "text":
-            previewDraw.ctx.font = draw.ctx.font;
+            previewDraw.ctx.font = drawer.ctx.font;
 
             previewDraw.clear();
             previewDraw.rectangle(lowestPosition.x, lowestPosition.y, getEventPos(event).x - lowestPosition.x, getEventPos(event).y - lowestPosition.y);
@@ -787,8 +346,8 @@ function handleMouseDownOrTouchStart(event) {
 
     console.log(event);
     if (currentDrawingMode !== "line") {
-        draw.moveTo(getEventPos(event).x, getEventPos(event).y);
-        draw.ctx.beginPath();
+        drawer.moveTo(getEventPos(event).x, getEventPos(event).y);
+        drawer.ctx.beginPath();
     }
 
     lowestPosition = getEventPos(event);
@@ -801,7 +360,7 @@ function handleMouseUpOrTouchEnd(event) {
     highestPosition = getEventPos(event);
 
     if (currentDrawingMode === "shape") {
-        drawShape(currentShapeToDraw, draw, { lowestPosition: lowestPosition, highestPosition: highestPosition }, currentShapeToDraw === "polygon" ? { sides: parseInt(document.querySelector("#polygon-sides-input").value) } : undefined);
+        drawShape(currentShapeToDraw, drawer, { lowestPosition: lowestPosition, highestPosition: highestPosition }, currentShapeToDraw === "polygon" ? { sides: parseInt(document.querySelector("#polygon-sides-input").value) } : undefined);
 
         drawingPreview.style.display = "none";
         previewDraw.clear();
@@ -815,7 +374,7 @@ function handleMouseUpOrTouchEnd(event) {
         drawingPreview.style.display = "none";
         previewDraw.clear();
 
-        drawText(draw);
+        drawText(drawer);
     }
 
     isDrawing = false;
@@ -841,7 +400,7 @@ function drawText(drawToDrawText) {
 /**
  * 
  * @param {String} shape Name of the shape to draw
- * @param {Draw} targetDraw The Draw to draw to
+ * @param {Drawer} targetDraw The Drawer to draw to
  * @param {{
  *      lowestPosition: {
  *          x: number,
@@ -887,13 +446,13 @@ function drawShape(shape, targetDraw, pos, values = {}) {
 
 for (let i = 0; i < document.querySelectorAll("button[data-shape]").length; i++) {
     const alternateCtx = document.querySelectorAll("button[data-shape]")[i].children[0].getContext("2d");
-    const temporaryDraw = new Draw(alternateCtx.canvas);
-    temporaryDraw.ctx.lineWidth = 1.5;
+    const temporaryDrawer = new Drawer(alternateCtx.canvas);
+    temporaryDrawer.ctx.lineWidth = 1.5;
 
     if (document.documentElement.classList.contains("dark-mode")) {
-        temporaryDraw.ctx.strokeStyle = "white";
+        temporaryDrawer.ctx.strokeStyle = "white";
     } else {
-        temporaryDraw.ctx.strokeStyle = "black";
+        temporaryDrawer.ctx.strokeStyle = "black";
     }
 
     if (alternateCtx.canvas.parentElement.getAttribute("data-shape") === "polygon") {
@@ -910,7 +469,7 @@ for (let i = 0; i < document.querySelectorAll("button[data-shape]").length; i++)
 
         alternateCtx.stroke();
     } else {
-        drawShape(alternateCtx.canvas.parentElement.getAttribute("data-shape"), temporaryDraw, { lowestPosition: { x: 0, y: 0 }, highestPosition: { x: 16, y: 16 } });
+        drawShape(alternateCtx.canvas.parentElement.getAttribute("data-shape"), temporaryDrawer, { lowestPosition: { x: 0, y: 0 }, highestPosition: { x: 16, y: 16 } });
     }
 
     if (alternateCtx.canvas.parentElement.getAttribute("data-shape") === "polygon") {
@@ -940,7 +499,7 @@ for (let i = 0; i < document.querySelectorAll("button[data-shape]").length; i++)
 for (let i = 0; i < document.querySelectorAll("button[data-type]").length; i++) {
     if (document.querySelectorAll("button[data-type]")[i].getAttribute("data-type") === "line") {
         document.querySelectorAll("button[data-type]")[i].addEventListener("click", function () {
-            draw.ctx.beginPath();
+            drawer.ctx.beginPath();
             currentDrawingMode = this.getAttribute("data-type");
             hideTextCaracteristics();
             showNewPathButton();
@@ -1018,15 +577,15 @@ for (let i = 0; i < document.querySelectorAll("button[data-type]").length; i++) 
         temporaryCtx.stroke();
         temporaryCtx.closePath();
     } else if (temporaryCtx.canvas.parentElement.getAttribute("data-type") === "clear") {
-        const temporaryDraw = new Draw(temporaryCtx.canvas);
+        const temporaryDrawer = new Drawer(temporaryCtx.canvas);
         temporaryCtx.lineWidth = 12;
         temporaryCtx.fillStyle = "crimson";
-        temporaryDraw.diamond(temporaryCtx.canvas.width / 4, temporaryCtx.canvas.height / 3 * 2, temporaryCtx.canvas.width / 3, temporaryCtx.canvas.height / 3, true);
+        temporaryDrawer.diamond(temporaryCtx.canvas.width / 4, temporaryCtx.canvas.height / 3 * 2, temporaryCtx.canvas.width / 3, temporaryCtx.canvas.height / 3, true);
         temporaryCtx.fillStyle = "cyan";
-        temporaryDraw.diamond(temporaryCtx.canvas.width / 4 * 1.66, temporaryCtx.canvas.height / 3 * 1.5, temporaryCtx.canvas.width / 3, temporaryCtx.canvas.height / 3, true);
+        temporaryDrawer.diamond(temporaryCtx.canvas.width / 4 * 1.66, temporaryCtx.canvas.height / 3 * 1.5, temporaryCtx.canvas.width / 3, temporaryCtx.canvas.height / 3, true);
     } else if (temporaryCtx.canvas.parentElement.getAttribute("data-type") === "text") {
-        const temporaryDraw = new Draw(temporaryCtx.canvas);
-        temporaryDraw.text("t", temporaryDraw.canvas.width / 7 * 2, -30, { fontSize: temporaryDraw.canvas.width, color: document.documentElement.classList.contains("dark-mode") ? "white" : "black" });
+        const temporaryDrawer = new Drawer(temporaryCtx.canvas);
+        temporaryDrawer.text("t", temporaryDrawer.canvas.width / 7 * 2, -30, { fontSize: temporaryDrawer.canvas.width, color: document.documentElement.classList.contains("dark-mode") ? "white" : "black" });
     }
 }
 lineWidthInput.addEventListener("input", function () {
@@ -1045,12 +604,12 @@ lineWidthInput.addEventListener("input", function () {
     }
 
     ctx.lineWidth = lineWidthInput.value;
-    draw.ctx.lineWidth = lineWidthInput.value;
+    drawer.ctx.lineWidth = lineWidthInput.value;
 })
 
 drawingColorInput.addEventListener("input", function () {
     ctx.strokeStyle = drawingColorInput.value;
-    draw.ctx.strokeStyle = drawingColorInput.value;
+    drawer.ctx.strokeStyle = drawingColorInput.value;
 })
 
 resizer.addEventListener("click", function () {
@@ -1080,7 +639,7 @@ closeCanvasBtn.addEventListener('click', function () {
 
     const saveBtn = document.createElement("button");
     const saveLinkWrraper = document.createElement("a");
-    saveLinkWrraper.href = "../index.html";
+    saveLinkWrraper.href = "/index.html";
     saveLinkWrraper.textContent = "Salvar";
     saveLinkWrraper.addEventListener('click', function () {
         canvas.toBlob((blob) => {
@@ -1091,7 +650,7 @@ closeCanvasBtn.addEventListener('click', function () {
 
     const doNotSaveBtn = document.createElement("button");
     const doNotSaveLinkWrraper = document.createElement("a");
-    doNotSaveLinkWrraper.href = "../index.html";
+    doNotSaveLinkWrraper.href = "/index.html";
     doNotSaveLinkWrraper.textContent = "Não salvar";
     doNotSaveBtn.appendChild(doNotSaveLinkWrraper);
