+++ src/lib/errors/DuplicateIDError.js
