+++ src/lib/polyfill/SharedWorkerPolyfill.js
